pyproject.toml
[project]
name = "n8ntelegramproof"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "uvicorn>=0.45.0",
]
